refactor(db): extract discovery pages to db/src/discovery.rs
